Feat: Add camera and photo library usage descriptions for iOS
This commit adds the `NSCameraUsageDescription` and `NSPhotoLibraryUsageDescription` keys to the `Info.plist` file for the iOS app. These descriptions are required by iOS to inform the user why the app needs access to the camera and photo library respectively.

Additionally, the `VerifierApp.swift` file was modified to ignore the safe area at the bottom edge for the `ContentView`.